kalabox/kalabox#1274 + kalabox/kalabox#1377: Cleanup sidebar tests and put in test for deleting apps.
